Create ME1115 Drive End.SLDPRT
Reverse modelled the ME1115 Drive End motor part.  The rest of the motor still needs to be reverse modelled so that the model matches the motor on the kart.  The main reason for this change was the mismatch between the drive shaft length and the model that was provided by the manufacturer.  Rear of the motor and the power hookups still need to be modeled at this time.