Replacement of supplied motor model
Remodeled the supplied motor model to include the extended length of the shaft that is found on the current kart.  Replaced in main kart assembly and fixed the motor mount assembly so that it was fully defined in the subassembly.  Had to slightly modify a few distances here and there to keep dimensions consistent between parts.  The weight of the motor is not correct and will need to be overridden.